requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.12
aioice==0.9.0
aiortc==1.11.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asyncpg==0.30.0
attrs==25.3.0
av==14.2.0
azure-core==1.33.0
azure-identity==1.21.0
backoff==2.2.1
banks==2.1.2
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.4
click==8.1.8
cloudevents==1.11.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.2
cssselect==1.3.0
cssselect2==0.8.0
dataclasses-json==0.6.7
dateutils==0.6.12
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
dotenv==0.9.9
durationpy==0.9
fastapi==0.115.9
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
flatbuffers==25.2.10
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.86.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.12.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.2.1
griffe==1.7.3
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
kubernetes==32.0.1
lazy-object-proxy==1.10.0
llama-cloud==0.1.19
llama-cloud-services==0.6.16
llama-index==0.12.33
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.33.post1
llama-index-embeddings-google-genai==0.1.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-openai==0.3.38
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-postgres==0.5.0
llama-parse==0.6.16
lxml==5.3.1
lxml_html_clean==0.4.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.3.0
mypy_extensions==1.1.0
narwhals==1.34.1
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.4
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.70.0
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opentelemetry-api==1.32.0
opentelemetry-exporter-otlp-proto-common==1.32.0
opentelemetry-exporter-otlp-proto-grpc==1.32.0
opentelemetry-instrumentation==0.53b0
opentelemetry-instrumentation-asgi==0.53b0
opentelemetry-instrumentation-fastapi==0.53b0
opentelemetry-proto==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
opentelemetry-util-http==0.53b0
orjson==3.10.16
oscrypto==1.3.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parse==1.20.2
pathable==0.4.4
pdfminer.six==20250327
pdfplumber==0.11.6
pgvector==0.4.0
pillow==11.1.0
platformdirs==4.3.7
posthog==3.24.0
prance==23.6.21.0
praw==7.8.1
prawcore==2.4.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.7
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybars4==0.9.13
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.10.1
pylibsrtp==0.11.0
PyMeta3==0.5.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
pypdf==5.4.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyTelegramBotAPI==4.26.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytube==15.0.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
ruamel.yaml==0.18.10
scipy==1.15.2
semantic-kernel==1.27.1
sgmllib3k==1.0.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.45.3
streamlit==1.44.1
striprtf==0.0.26
svglib==1.5.1
sympy==1.13.3
telebot==0.0.5
tenacity==9.1.2
tiktoken==0.9.0
tinycss2==1.4.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
update-checker==0.18.0
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.1
wikipedia==1.4.0
wrapt==1.17.2
xhtml2pdf==0.2.17
yarl==1.18.3
youtube-transcript-api==1.0.3
zipp==3.21.0
